Refactor imports in sections to use ~/graphql/data/queries for improved organization and clarity in query management
